relation ManyToMany and  creating controllers

+++ src/main/java/com/springboot/sportscomplex/models/dto/SubscriptionDTO.java
@@ -1,6 +1,5 @@
 package com.springboot.sportscomplex.models.dto;
 
-import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
 
 import java.io.Serializable;
@@ -8,10 +7,7 @@ import java.io.Serializable;
 @Data
 public class SubscriptionDTO implements Serializable {
 
-//    @NotBlank
-    private int priceGymSubscription;
-//    @NotBlank
-    private int priceSwimmingPoolSubscription;
-//    @NotBlank
-    private int priceSaunaSubscription;
+    private double priceGymSubscription;
+    private double priceSwimmingPoolSubscription;
+    private double priceSaunaSubscription;
 }
